Working on getting msd reads to work
I can't seem to figure out if there is any issue with bochs at this
point. It should be working, but on bochs it keeps defering my
in-transactions

+++ kernel/arch/x86/x32/idt.c
@@ -83,7 +83,7 @@ void IdtInitialize(void)
 	InterruptInstallDefaultGates();
 
 	/* Override default gate with syscall */
-	IdtInstallDescriptor(INTERRUPT_SYSCALL, (uint32_t)syscall_entry, 
+	IdtInstallDescriptor(INTERRUPT_SYSCALL, (uintptr_t)syscall_entry, 
 		GDT_KCODE_SEGMENT, IDT_RING3 | IDT_PRESENT | IDT_TRAP_GATE32);
 
 	/* Reload GDT */
